Reorganize modules around the tenant mapping and bootstrapping
