--- rbac/management/tenant/__init__.py
@@ -1,1 +1,0 @@
-"""Tenant management business logic."""
