+++ rbac/management/tenant_mapping/__init__.py
@@ -1,0 +1,1 @@
+"""Tenant Mapping model."""
